[DEV-8580] more refactors
