--- packages/chart/src/components/Sankey/ColumnList.tsx
@@ -1,17 +1,0 @@
-interface ColumnData {
-  label: string
-  value: string
-  additional_info: string
-}
-
-const ColumnList: React.FC<{ columnData: ColumnData[] }> = ({ columnData }) => {
-  return (
-    <ul>
-      {columnData?.map((entry, index) => {
-        return <li key={index}>{entry}</li>
-      })}
-    </ul>
-  )
-}
-
-export default ColumnList
